feat: Refactor `dsel-module-submodules` to use an alist
This change modifies the `submodules` slot of `dsel-module` from a list of modules to an alist of `(name . module)` pairs.

This allows for named access to submodules and enables the new `dsel-module-get-submodule` generic function. The `dsel-module-deepcopy` and `dsel-collect-predictors` methods have been updated to reflect this change. A new test case for `dsel-module-get-submodule` has also been added.